# service-main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "starting point server",
        "title": "service-main",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/msg": {
            "post": {
                "description": "put msg in write service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "WriteMsg",
                "operationId": "write-msg",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/msg/{id}": {
            "get": {
                "description": "get msg from read service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "msg"
                ],
                "summary": "ReadMsg",
                "operationId": "get-msg",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Message": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "msg": {
                    "type": "string",
                    "example": "any text"
                }
            }
        }
    }
}